requirements.txt
# Core dependencies
structlog>=23.0.0
click>=8.1.0
requests>=2.28.0

# Data analysis and similarity detection
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.0

# LLM API libraries (choose what you need)
openai>=1.0.0
anthropic>=0.20.0
langchain>=0.1.0
google-generativeai>=0.3.0
litellm>=1.0.0

# Web dashboard (optional)
fastapi>=0.100.0
uvicorn>=0.23.0
websockets>=11.0
jinja2>=3.1.0

# Utilities
python-dotenv>=1.0.0
